requirements.txt
gradio>=4.0.0
openai>=1.0.0
langchain>=0.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
gradio>=4.0.0
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
 numpy>=1.24.0
 pandas>=2.0.0
 PyGithub>=2.1.0
 boto3>=1.28.0